reatment of problem with the tracing , missing links
One problem we faced was that we did not draw all points of ways. This was poorly secured buildings or of not fully drawn roads (there were holes )

To remedy this, I tested the storage whether we parsed all of the components , the problem did not come from there.
The problem was not in retrieving information stored. The problem was that we drew the ways with the wrong number of node , it lacked an argument but was present in the table.

The problem is corrected